move nullability check test to ConditionalExpressionSuite and add withNullableElseValue
