+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/optimizer/SimplifyConditionalSuite.scala
@@ -293,22 +293,4 @@ class SimplifyConditionalSuite extends PlanTest with ExpressionEvalHelper {
         Some(Literal.create(null, IntegerType))),
       CaseWhen((GreaterThan($"a", 1), Literal.create(1, IntegerType)) :: Nil))
   }
-
-  test("accurate") {
-    val nullLiteral = Literal.create(null, BooleanType)
-    val noElseValue = CaseWhen(normalBranch :: trueBranch :: Nil, None)
-    assert(!noElseValue.nullable)
-    val withElseValue = CaseWhen(normalBranch :: trueBranch :: Nil, Some(Literal(1)))
-    assert(!withElseValue.nullable)
-    val firstTrueNonNullableSecondTrueNullable = CaseWhen(trueBranch ::
-      (TrueLiteral, nullLiteral) :: Nil, None)
-    assert(!firstTrueNonNullableSecondTrueNullable.nullable)
-    val firstTrueNullableSecondTrueNonNullable = CaseWhen((TrueLiteral, nullLiteral) ::
-      trueBranch :: Nil, None)
-    assert(firstTrueNullableSecondTrueNonNullable.nullable)
-    val hasNullInNotTrueBranch = CaseWhen(trueBranch :: (FalseLiteral, nullLiteral) :: Nil, None)
-    assert(!hasNullInNotTrueBranch.nullable)
-    val noTrueBranch = CaseWhen(normalBranch :: Nil, Literal(1))
-    assert(!noTrueBranch.nullable)
-  }
 }
